[Feat: #147219801] Implement Profile Model and Test models - Add Profile model for users - Improve test structure by refactoring test for every model to a module - Add tests for profile - Migrate new profile model - Update relationship among models   - User and Profile -> owner and profile_id   - Item and Bucket -> asset_id   - Item and Bucket and Profile -> profile_id
