+++ app/models/bucketlist.py
@@ -1,8 +1,8 @@
 """This module contains Modlels for Item and Bucketlist."""
 
-from sqlalchemy.orm.collections import attribute_mapped_collection
-
 from app.base import BaseModel, database
+from sqlalchemy.ext.declarative import declared_attr
+from sqlalchemy.orm.collections import attribute_mapped_collection
 
 
 class Base(database.Model, BaseModel):
@@ -14,7 +14,8 @@ class Base(database.Model, BaseModel):
     __abstract__ = True
     id = database.Column(database.Integer, primary_key=True)
     asset_id = database.Column(database.Integer)
-    name = database.Column(database.String(128), nullable=False, unique=True)
+    name = database.Column(database.String(128), nullable=False, unique=False)
+
     date_created = database.Column(
         database.DateTime, default=database.func.now())
 
@@ -23,8 +24,23 @@ class Base(database.Model, BaseModel):
         default=database.func.now(),
         onupdate=database.func.now())
 
+    @declared_attr
+    def profile_id(self):
+        """Profile Id, links bucket/item to a profile."""
+        return database.Column(database.Integer,
+                               database.ForeignKey('profiles.id'))
+
+    def to_dict(self):
+        """Give a Dictionary represantation for the model."""
+        return {
+            "name": self.name,
+            "id": self.asset_id,
+            "date created": str(self.date_created),
+            "date modified": str(self.date_modified)
+        }
+
     @classmethod
-    def get_object(cls, name=None, id=None, asset_id=None):
+    def get_object(cls, name=None, id=None, asset_id=None, profile_id=None):
         """Query database using the specified param as a filter.
 
         Args:
@@ -36,11 +52,13 @@ class Base(database.Model, BaseModel):
             An instance of the class passed in cls if succefull else None
         """
         if name:
-            return cls.query.filter_by(name=name).first()
+            return cls.query.filter_by(name=name).filter_by(
+                profile_id=profile_id).first()
         elif id or id == 0:
             return cls.query.filter_by(id=id).first()
-        elif asset_id or asset_id == 0:
-            return cls.query.filter_by(asset_id=asset_id).first()
+        elif (asset_id or asset_id == 0) and profile_id:
+            return cls.query.filter_by(asset_id=asset_id).filter_by(
+                profile_id=profile_id).first()
         else:
             return cls.query.all()
 
@@ -51,29 +69,33 @@ class Bucket(Base):
     __tablename__ = "buckets"
     items = database.relationship(
         "Item",
-        collection_class=attribute_mapped_collection('name'),
+        collection_class=attribute_mapped_collection('asset_id'),
         cascade="all, delete-orphan")
 
     created_by = database.Column(database.String(128))
-
-    profile_id = database.Column(database.Integer,
-                                 database.ForeignKey('profiles.id'))
 
     def __init__(self, name=""):
         """Initialize a bucket with it's name."""
         self.name = name
 
     def to_dict(self):
+        """Give a Dictionary represantation for the model."""
+        global index
+        index = -1
 
-        return dict(
-            name=self.name,
-            profile_id=self.profile_id,
-            id=self.asset_id,
-            items={key: self.items[key].to_dict()
-                   for key in self.items},
-            created_by=self.created_by,
-            date_created=str(self.date_created),
-            date_modified=str(self.date_modified))
+        def plus_one():
+            global index
+            index = index + 1
+            return index
+
+        base_dict = Base.to_dict(self)
+        base_dict.update({
+            "items":
+            {plus_one(): self.items[key].to_dict()
+             for key in self.items},
+            "created by": self.created_by
+        })
+        return base_dict
 
     @classmethod
     def get_buckets(cls):
@@ -101,15 +123,12 @@ class Bucket(Base):
         """
         bucket = None
         if name:
-            bucket = Bucket.get_object(name=name)
+            bucket = Bucket.get_object(name=name, profile_id=profile_id)
         elif id or id == 0:
             bucket = Bucket.get_object(id=id)
         elif asset_id or asset_id == 0:
-            bucket = Bucket.get_object(asset_id=asset_id)
-
-        if (profile_id or profile_id == 0) and bucket:
-            if not (profile_id == bucket.profile_id):
-                bucket = False
+            bucket = Bucket.get_object(
+                asset_id=asset_id, profile_id=profile_id)
         return bucket
 
 
@@ -129,14 +148,13 @@ class Item(Base):
         self.description = description
 
     def to_dict(self):
-        return dict(
-            name=self.name,
-            description=self.description,
-            done=self.done,
-            bucket_id=self.bucket_id,
-            id=self.asset_id,
-            date_created=str(self.date_created),
-            date_modified=str(self.date_modified))
+        """Give a Dictionary represantation for the model."""
+        base_dict = Base.to_dict(self)
+        base_dict.update({
+            "description": self.description,
+            "done": self.done,
+        })
+        return base_dict
 
     @classmethod
     def get_item(cls, name=None, id=None, asset_id=None, profile_id=None):
@@ -152,13 +170,9 @@ class Item(Base):
         """
         item = None
         if name:
-            item = cls.get_object(name=name)
+            item = cls.get_object(name=name, profile_id=profile_id)
         elif id or id == 0:
             item = cls.get_object(id=id)
         elif asset_id or asset_id == 0:
-            item = cls.get_object(asset_id=asset_id)
-
-        if (profile_id or profile_id == 0) and item:
-            if not (profile_id == item.bucket_id):
-                item = False
+            item = cls.get_object(asset_id=asset_id, profile_id=profile_id)
         return item
